refactor: extract shared components, DRY provider labels, split large file
- Extract providerLabel utility to lib/providerLabels.ts — eliminates
  3x duplication across ConnectedAccountsSection, ConnectorCredentials,
  and OAuthProviderSection
- Split OAuthCredentialRow into dedicated file (was inline in 420-line
  ConnectorCredentialsSection)
- Extract StoredCredentialList component to DRY up credential listing
  markup shared between OAuthCredentialRow and CredentialRow
- Integrate OAuthFailReason check into main's resolveCredentialsWithFallback
  to prevent silent credential masking on broken OAuth connections
